Rename settings.yaml to goosebit.yaml, and check multiple places for settings.
Use can now pass `GOOSEBIT_SETTINGS` to customize settings location, as well as goosebit auto checking `/etc/goosebit` if it exists, the root of the project directory, and the current working directory when it is run.
This should allow for more optionality when loading settings files.

Renaming settings to `goosebit.yaml` is mainly to specify what the file is when running from CWD, it should be distinct from other files.

Fixes: #103