Use `model_validate` built into pydantic to parse DB models
`pydantic` models have a deprecated `from_orm` method, which is designed to parse DB models into the schema.
This was superseded by `model_validate`, which does the same thing when the schema config `from_attributes` value is true.
Migrating to using this simplifies the method of parsing DB models, and will eventually allow the schemas to be used internally instead of the DB models.