Remove `poll_time_updating`
SWUpdate has its own thread that regularly sends log updates
while update is being installed. So no need to adjust the
poll time.

Only impact of a lower poll time is that SWUpdate checks more
often if installation should be cancelled. As this functionality
is not offered by gooseBit the custom poll time can be removed.